rename Int2IntMap to Int2IntOpenHashMap and move it out of MapPalette, simplify tableSizeFor
